% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celegance.R
\docType{data}
\name{Celegance}
\alias{Celegance}
\title{Celegance}
\format{
a character vector
}
\usage{
data("Celegance")
}
\description{
A sample collection of 49 UNIPROT_AC of Celegance species.
}
\examples{
data("Celegance")
print(Celegance)
}
\references{
http://www.uniprot.org/
}
\keyword{datasets}
