% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FruitFly.R
\docType{data}
\name{FruitFly}
\alias{FruitFly}
\title{Fruit Fly}
\format{a character vector}
\usage{
data("FruitFly")
}
\description{
A sample collection of 56 UNIPROT_AC of Fruit Fly species.
}
\examples{
data("FruitFly")
print(FruitFly)
}
\references{
http://www.uniprot.org/
}
\keyword{datasets}
