% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H.sapiens.R
\docType{data}
\name{H.sapiens}
\alias{H.sapiens}
\title{Homo sapiens (Human)}
\format{
a character vector
}
\usage{
data("H.sapiens")
}
\description{
A sample collection of 76 UNIPROT_AC of Homo sapiens species.
}
\examples{
data("H.sapiens")
print(H.sapiens)
}
\references{
http://www.uniprot.org/
}
\keyword{datasets}
