% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMMAN.R
\name{IMMAN}
\alias{IMMAN}
\title{Interlog protein network reconstruction by Mapping and Mining ANalysis}
\usage{
IMMAN(ProteinLists, fileNames = NULL, Species_IDs, identityU,
  substitutionMatrix, gapOpening, gapExtension, BestHit, coverage,
  NetworkShrinkage, score_threshold, STRINGversion, InputDirectory = getwd())
}
\arguments{
\item{ProteinLists}{a list in which each element contains protein names of a species as a character vector.
If it was NULL then the protein lists file name should be addressed in fileNames parameter.}

\item{fileNames}{a character vector, containing names of text files containing protein list for each species.
The protein list of each species must be in a column without header and rownames in seperate ".txt" files.
The ProteinLists argument should be include at least two text file names addressing the protein list of each species
which are in UniProt accession IDs format.}

\item{Species_IDs}{a numeric vector; taxonomy ID for each organism which are provided in fileNames}

\item{identityU}{numeric; value for selecting proteins whose alignment score is greater or equal than identityU}

\item{substitutionMatrix}{a scoring substitution matrix to be used for alignment setting.}

\item{gapOpening}{numeric; indicating the cost for opening a gap in the alignment}

\item{gapExtension}{The incremental cost incurred along the length of the gap in the alignment}

\item{BestHit}{logical; if TRUE describes a pair protein sequence among two different species
which is the reciprocal best hit in sequence similarity analysis, whilst,
if it is FALSE, indicates a nonreciprocal best hit}

\item{coverage}{Number of connected proteins pairs in each Ortholog Protein Set (OPS) pair
(termed as “coverage”) to
reconstruct an edge of OPS pair in the IPN (Interlog Protein Network)}

\item{NetworkShrinkage}{logical; if TRUE OPSs that are similar to each other would be merged.}

\item{score_threshold}{numeric; STRINGdb score for protein protein interaction (PPI) selection in STRING database}

\item{STRINGversion}{character; indicating which version of STRING database should program
search in for the score of PPIs.}

\item{InputDirectory}{By default is getwd(). You can set this parameter to indicate where
the downloaded file from STRING should be saved.}
}
\value{
a list contaning four elements:

IPNEdges : data.frame; Edges of resulted interlog protein network.

IPNNodes : data.frame; Nodes of resulted interlog protein network.
Each node represents an OPS which is a set of ortholog proteins.

Network : list; Retrived PPINs of each input species.

maps : list; It includes data.frames indicating STRING_id data base matched to their corresponding
UNIPROT_AC. The number of data.frames is according to the the number of species.

IPN : an igraph object representing the interlog protein network.
}
\description{
A function for reconstructing Interlog Protein Network (IPN) integrated
from  Protein-protein Interaction Networks (PPIN) from different species. Users can overlay different
PPINs to mine conserved common network between diverse species. It helps to retrieve
IPN with different degrees of conservation to have better protein function
prediction and PPIN analysis.
}
\details{
This function is specified to retrieve the interlog protein network shared across diverse species.
For this aim, first we exceed orthology relationships among sequences from various species by iterating
over any pair of input species, and using the Needleman-Wunsch alignment algorithm and a best reciprocal hit strategy to
reach the orthologues through all versus all pairwise cross-species alignments. From orthology
assignment, we derive Orthologous Protein sets (OPSs), an assortment of clusters of orthologues
(maximum one per species) which will conform the nodes of the so-called Interolog Protein Network (IPN).
We exceed n species-specific interlog protein networks from STRING database, where each node maps to a
single OPS in the IPN, and distinguish the edges of the outcome IPN by choosing only edges linking nodes
in the IPN that also are linked in at least k species-specific networks (where 'k' is set as a parameter).

A scoring system is used by the alignment process, which can be described as a set of values specified for quantifying
the likelihood of one residue that has been substituted by another in an alignment. The scoring systems used by
alignment procedure is called a substitution matrix and it can be achieved from statistical analysis of residue substitution
data from sets of reputable alignments of highly relevent sequences.

Using identityU value which ranges from 0 to 100, user would be able to specialize how the IPNs should be larger or not.
As the value of identityU gets higher, the algorithm will find much similar orthologs and vice versa.

We used gapOpening and gapExtension arguments to figure numeric values of ortholog proteins.
For matching alignments of proteins if we skip a protein, gapOpening argument would be incremented.
The smaller the amount of gap, protein alignements are more similar to each other.

The score_threshold argument is specified for evaluating the similarity values between two proteins
in substitutionMatrix. It differs from 0 to 100, however, the common use ranges from 25 to 30.

The transference of interactionn among orthologs of different species called the interlog approach.

We used Besthit argument to reach proteins which has the most similarity in all versus all protein alignment.

If an interaction was exist between each pair of proteins of OPSs, an edge would be linked in the IPN. The coverage_threshold
specifies the number of interactions that are exist among pair of proteins of OPSs. It differs from 1 to number of species.
As much as the value of coverage_threshold was high, the final IPN would be more robust and usually smaller.

NetworkShrinkage argument determine whether two similar OPSs which have ortholog proteins in common should be merged or not.
If it was TRUE the resulting IPN would be smaller.
}
\examples{
data(H.sapiens)
data(R.norvegicus)
data(Celegance)
data(FruitFly)

ProteinLists = list(as.character(H.sapiens$V1), as.character(R.norvegicus$V1),
 as.character(Celegance$V1), as.character(FruitFly$V1))

List1_Species_ID = 9606  # taxonomy ID List1 Homo sapiens
List2_Species_ID = 10116 # taxonomy ID List2 Rat
List3_Species_ID = 6239  # taxonomy ID List3 Celegans
List4_Species_ID = 7227  # taxonomy ID List4 FruitFly

Species_IDs  = c(List1_Species_ID, List2_Species_ID, List3_Species_ID, List4_Species_ID)

# Set the parameters to run the analysis.
# Here is a description of parameters in IMMAN. If you need more
# information you can refer to the paper.
# identityU: Cut off value for selecting proteins whose alignment score
# is greater or equal than identityU
# substitutionMatrix: Which scoring matrix to be used for alignment
# setting gapOpening and gapExtension for alignment purposes
# For NetworkShrinkage, coverage, and BestHit refer to paper
# score_threshold:
# STRINGversion: Indicated which version of STRING database should
# program search in for the score of PPIs.

identityU = 30
substitutionMatrix = "BLOSUM62"
gapOpening = -8
gapExtension = -8
NetworkShrinkage = FALSE
coverage = 1
BestHit = TRUE
score_threshold = 400
STRINGversion="10"

# Run the IMMAN function for the parameters
output = IMMAN(ProteinLists, fileNames=NULL, Species_IDs,
              identityU, substitutionMatrix,
              gapOpening, gapExtension, BestHit,
              coverage, NetworkShrinkage,
              score_threshold, STRINGversion,
              InputDirectory = getwd())

output$IPNEdges
output$IPNNodes
output$Networks
output$Networks[[1]]
output$maps
output$maps[[2]]

}
\seealso{
\code{\link[Biostrings]{PairwiseAlignments}}
}
\author{
Payman Nickchi, Abdollah Safari, Minoo Ashtiani, Mohieddin Jafari
}
