% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.norvegicus.R
\docType{data}
\name{R.norvegicus}
\alias{R.norvegicus}
\title{Rattus norvegicus (Rat)}
\format{a character vector}
\usage{
data("R.norvegicus")
}
\description{
A sample collection of 70 UNIPROT_AC of Rattus norvegicus species.
}
\examples{
data("R.norvegicus")
print(R.norvegicus)
}
\references{
http://www.uniprot.org/
}
\keyword{datasets}
